refactor: move parsing of any metadata into utils
