+++ packages/ui/package.json
@@ -22,7 +22,6 @@
     "dayjs": "^1.11.13",
     "helia": "^5.2.0",
     "highlight.js": "^11.11.1",
-    "js-yaml": "^4.1.1",
     "leaflet": "^1.9.4",
     "lucide-vue-next": "^0.544.0",
     "multiformats": "^13.3.1",
