+++ packages/utils/package.json
@@ -15,11 +15,12 @@
     "typescript": "^5.6.3"
   },
   "dependencies": {
+    "@helia/ipns": "^8.1.0",
+    "@helia/verified-fetch": "^2.4.0",
     "dayjs": "^1.11.13",
     "helia": "^5.2.0",
-    "zarrita": "^0.4.0-next.19",
-    "@helia/verified-fetch": "^2.4.0",
-    "@helia/ipns": "^8.1.0",
-    "ipfs-unixfs-exporter": "^13.6.1"
+    "ipfs-unixfs-exporter": "^13.6.1",
+    "js-yaml": "^4.1.1",
+    "zarrita": "^0.4.0-next.19"
   }
 }
